requirements.txt
# requirements.txt

streamlit
pandas
requests
textblob
scikit-learn
matplotlib
pymongo[srv]
python-dotenv
bcrypt